backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Backend do app emailClassifier"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi",
    "uvicorn[standard]",
    "python-multipart",
    "pymupdf", 
    "google-generativeai",
    "python-dotenv",
]

[project.optional-dependencies]
dev = [
    "ruff"
]
